+++ frontend/javascripts/oxalis/model/sagas/flycam_info_cache_saga.ts
